trading-agent/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
python-dotenv==1.0.0
anthropic==0.34.0
yfinance==0.2.43
pandas==2.2.2
numpy==2.1.0
pandas-ta==0.3.14b
sqlalchemy==2.0.35
aiohttp==3.10.5
httpx==0.27.2
